+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 – Address properties are "" if empty (handy when comparing strings)
 - replace invalid characters when loading (the way the e-mail clients behave)
 - headers truly case-insensitive when loading
+- `.from_()` and `.sender()` truly return an `Address` even if not set (and then the Address is empty)
 
 ## 1.4 (2021-01-23)
 - exposed Address